Finished the 'Laravel from Scratch' YouTube tutorial... Added user authentication, access control, and photo uploading

+++ app/Http/Controllers/PostsController.php
@@ -3,10 +3,22 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 
 class PostsController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -43,13 +55,38 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle File Upload
+        if($request->hasFile('cover_image')){
+            //Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+
+            //Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+
+            //Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            //Filename to store
+            $filenameToStore = $filename.'_'.time().'.'.$extension;
+
+            //Upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $filenameToStore);
+        }
+        else {
+            $filenameToStore = 'noimage.jpg';
+        }
+        
 
         //Add post to DB
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $filenameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post created');
@@ -79,10 +116,16 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+        $post = Post::find($id);
         $data = array(
             'title' => "Posts",
-            'post' => Post::find($id)
+            'post' => $post
         );
+
+        //Check for correct user
+        if(auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
 
         return view('posts.edit')->with($data);
     }
@@ -101,10 +144,37 @@ class PostsController extends Controller
             'body' => 'required'
         ]);
 
-        //Add post to DB
+        //Check for correct user
+        if(auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+        // Handle File Upload
+        if($request->hasFile('cover_image')){
+            //Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+
+            //Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+
+            //Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            //Filename to store
+            $filenameToStore = $filename.'_'.time().'.'.$extension;
+
+            //Upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $filenameToStore);
+        }
+
+        //Add updated post to DB
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $filenameToStore;
+        }
+
         $post->save();
 
         return redirect('/posts')->with('success', 'Post updated');
@@ -119,6 +189,17 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        //Check for correct user
+        if(auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+        if($post->cover_image != 'noimage.jpg') {
+            //Delete the image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post removed');
     }
